restructure utils based on PR comments
